Reorganise project structure for ui
