+++ src/main/java/seedu/address/ui/MainWindow.java
@@ -12,7 +12,8 @@ import seedu.address.logic.Logic;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.ui.servicetab.ServiceListPanel;
+import seedu.address.ui.clientspanel.ClientListPanel;
+import seedu.address.ui.servicespanel.ServiceListPanel;
 
 /**
  * The Main Window. Provides the basic application layout containing
