+++ src/main/java/seedu/address/ui/clientspanel/ClientCard.java
@@ -1,13 +1,14 @@
-package seedu.address.ui;
-
-import java.util.Comparator;
+package seedu.address.ui.clientspanel;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
 import javafx.scene.layout.FlowPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
+
+import java.util.Comparator;
 import seedu.address.model.client.Client;
+import seedu.address.ui.UiPart;
 
 /**
  * An UI component that displays information of a {@code Client}.
