+++ src/main/java/seedu/address/ui/servicespanel/ServiceCard.java
@@ -1,4 +1,4 @@
-package seedu.address.ui.servicetab;
+package seedu.address.ui.servicespanel;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
