+++ src/main/resources/view/servicespanel/ServiceListCard.fxml
